Move SessionSerializer to controller
