+++ apps/helix_account/config/config.exs
@@ -16,6 +16,6 @@ config :guardian, Guardian,
   ttl: {1, :days},
   allowed_algos: ["HS512"],
   allowed_drift: 2_000,
-  serializer: Helix.Account.Model.SessionSerializer
+  serializer: Helix.Account.Controller.Session
 
 import_config "#{Mix.env}.exs"
